requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
asyncpg==0.30.0
attrs==25.3.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.35.0
botocore==1.35.0
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.2
certifi==2025.4.26
certipy==0.1.3
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==43.0.0
curl_cffi==0.10.0
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.24.0
distro==1.8.0
executing==2.0.1
fastapi==0.115.12
fastjsonschema==2.19.1
feedparser==6.0.11
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
inflection==0.5.1
ipykernel==6.29.2
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.17
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterhub==5.0.0
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kafka-python==2.2.11
kiwisolver==1.4.5
lxml==5.2.2
Mako==1.3.5
mangum==0.19.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mutagen==1.47.0
nbclient==0.9.0
nbconvert==7.16.4
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.2
oauthlib==3.2.2
openai==1.4.0
openapi==1.1.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.3
peewee==3.17.6
pg8000==1.31.2
pika==1.3.2
Pillow==9.3.0
platformdirs==4.2.0
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf>=4.21.0,<5.0.0
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyKakao==0.0.6
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytimeparse==1.1.8
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
scramp==1.4.5
seaborn==0.13.2
selenium==4.33.0
Send2Trash==1.8.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.46.2
sympy==1.12
terminado==0.18.0
tinycss2==1.2.1
torch==2.1.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.30.0
trio-websocket==0.12.2
twelvedata==1.2.12
types-python-dateutil==2.8.19.20240106
typing_extensions==4.13.2
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.10
wsproto==1.2.0
yarl==1.9.3
yfinance==0.2.28
youtube-search==2.1.2
yt-dlp==2023.11.16
zipp==3.17.0
